requirements.txt
Flask==1.1.2
line-bot-sdk==1.16.0

requests==2.20.1
Pillow==5.3.0
